Separate Aux_Record_ClusterCenter to different file
